Corrected error fron Enrollment and Student

+++ src/main/java/com/example/proyectogrupal/repositories/EnrollmentRepository.java
@@ -1,6 +1,5 @@
 package com.example.proyectogrupal.repositories;
 
-import com.example.proyectogrupal.entity.Course;
 import com.example.proyectogrupal.entity.Enrollment;
 import jakarta.persistence.EntityManager;
 import org.springframework.stereotype.Repository;
@@ -12,8 +11,7 @@ import java.util.Map;
 
 @Repository
 public class EnrollmentRepository implements EnrollmentRepositoryContract{
-    private final Map<Long, Enrollment> enrollment = new HashMap<>();
-
+    private final Map<Long, Enrollment> enrollments = new HashMap<>();
     private EntityManager entityManager;
 
     public EnrollmentRepository(EntityManager entityManager) {
@@ -22,17 +20,27 @@ public class EnrollmentRepository implements EnrollmentRepositoryContract{
 
     @Override
     public List<Enrollment> allEnrollment() {
-        return new ArrayList<>(enrollment.values());
+        return new ArrayList<>(enrollments.values());
     }
 
     @Override
-    public void findByIdEnrollment(Integer id) {
-        entityManager.find(Enrollment.class,id);
+    public Enrollment findByIdEnrollment(Integer id) {
+        if(id == null) {
+            return null;
+        }
+        return entityManager.find(Enrollment.class, id);
     }
 
     @Override
-    public void save(Enrollment enrollment) {
-        entityManager.persist(enrollment);
+    public Enrollment save(Enrollment enrollment) {
+        if (enrollment.getIdEnrollment() == null) {
+
+            entityManager.persist(enrollment);
+            return enrollment;
+
+        } else {
+            return entityManager.merge(enrollment);
+        }
     }
 
     @Override
@@ -42,6 +50,8 @@ public class EnrollmentRepository implements EnrollmentRepositoryContract{
 
     @Override
     public void delete(Integer id) {
-        entityManager.remove(id);
+        if (id != null) {
+            entityManager.remove(id);
+        }
     }
 }
